Overhaul to fix list popup overflow limiting problems.
This change moves the list popup out of the multiselect container and into the doc body, replacing
its relative positining with absolute positioning based on the location of the monitor. Formerly,
if a multiselect list appeared near the bottom of a container div, it would either enlarge the
div or be truncated when opened. This ensures that the list isn't constrained by the element
containing the monitor. It also clears up redundant argument passing, moving the list and monitor
variables into the object data to simplify the interface.